+++ src/internalnode.h
@@ -20,21 +20,21 @@ public:
     size_t size() const override;
     size_t minSize() const override;
     size_t maxSize() const override;
-    KeyType keyAt(int index) const;
-    void setKeyAt( int index, KeyType key );
+    KeyType keyAt( size_t index) const;
+    void setKeyAt( size_t index, KeyType key );
     Node* firstChild() const;
     void populateNewRoot( Node* oldNode, KeyType newKey, Node* newNode );
     size_t insertNodeAfter( Node* oldNode, KeyType newKey, Node* newNode );
-    void remove( int index );
+    void remove( size_t index );
     Node* removeAndReturnOnlyChild();
     KeyType replaceAndReturnFirstKey();
     void moveHalfTo( InternalNode* recipient );
-    void moveAllTo( InternalNode* recipient, int indexInParent );
+    void moveAllTo( InternalNode* recipient, size_t indexInParent );
     void moveFirstToEndOf( InternalNode* recipient );
-    void moveLastToFrontOf( InternalNode* recipient, int parentIndex );
+    void moveLastToFrontOf( InternalNode* recipient, size_t parentIndex );
     Node* lookup( KeyType key ) const;
-    int nodeIndex( Node* node ) const;
-    Node* neighbor( int index ) const;
+    size_t nodeIndex( Node* node ) const;
+    Node* neighbor( size_t index ) const;
     std::string toString( bool verbose = false ) const override;
     void queueUpChildren( std::queue< Node* >* queue );
 
@@ -42,7 +42,7 @@ private:
     void copyHalfFrom( std::vector< MappingType >& mappings );
     void copyAllFrom( std::vector< MappingType >& mappings );
     void copyLastFrom( MappingType pair );
-    void copyFirstFrom( MappingType pair, int parentIndex );
+    void copyFirstFrom( MappingType pair, size_t parentIndex );
 
 private:
     std::vector<MappingType> m_mappings;
